Azbox miniME onlineflash/multiboot

+++ lib/python/Plugins/SystemPlugins/SoftwareManager/Flash_online.py
@@ -80,10 +80,12 @@ class FlashOnline(Screen):
 		self.selection = 0
 		self.devrootfs = "/dev/mmcblk0p3"
 		self.multi = 1
+		self.AZrootfsMTD = None
+		self.AZkernelMTD = None
 		self.list = self.list_files("/boot")
 
 		Screen.setTitle(self, _("Flash On the Fly"))
-		if SystemInfo["HaveMultiBoot"]:
+		if SystemInfo["HaveMultiBoot"] and not getBoxType().startswith('azboxm'):
 			self["key_yellow"] = StaticText(_("STARTUP"))
 		else:
 			self["key_yellow"] = StaticText("")
@@ -102,10 +104,32 @@ class FlashOnline(Screen):
 			"cancel": self.quit,
 		}, -2)
 		if SystemInfo["HaveMultiBoot"]:
-			self.multi = self.read_startup("/boot/" + self.list[self.selection]).split(".",1)[1].split(" ",1)[0]
-			self.multi = self.multi[-1:]
-			print "[Flash Online] MULTI:",self.multi
-
+			if not getBoxType().startswith('azboxm'):
+				self.multi = self.read_startup("/boot/" + self.list[self.selection]).split(".",1)[1].split(" ",1)[0]
+				self.multi = self.multi[-1:]
+				print "[Flash Online] MULTI:",self.multi
+			else:
+				self.AZkernelMTD = self.get_current_azboot()
+				self.AZrootfsMTD = self.AZkernelMTD + 1
+				print "[Flash_online: AZkernelMTD=]", self.AZkernelMTD
+				print "[Flash_online: AZrootfsMTD=]", self.AZrootfsMTD
+				
+	def get_current_azboot(self):
+		if not os.path.isfile('/boot/.boot'):
+			os.system("mount -t jffs2 mtd2 /boot/")
+		if os.path.isfile('/boot/.boot'):
+			f=open('/boot/.boot','r')
+			current_bootnum = int(f.read())
+			f.close()		
+		#os.system("umount -fl /boot")
+		print "[Flash_online: current Azboot=]", current_bootnum
+		if current_bootnum == 0:
+			return 2
+		elif current_bootnum == 2:
+			return 4
+		elif current_bootnum == 3:
+			return 6
+		
 	def check_hdd(self):
 		if not os.path.exists("/media/hdd"):
 			self.session.open(MessageBox, _("No /hdd found !!\nPlease make sure you have a HDD mounted.\n\nExit plugin."), type = MessageBox.TYPE_ERROR)
@@ -139,18 +163,18 @@ class FlashOnline(Screen):
 		
 	def blue(self):
 		if self.check_hdd():
-			self.session.open(doFlashImage, online = False, list=self.list[self.selection], multi=self.multi, devrootfs=self.devrootfs)
+			self.session.open(doFlashImage, online = False, list=self.list[self.selection], multi=self.multi, AZrootfsMTD = self.AZrootfsMTD, AZkernelMTD = self.AZkernelMTD, devrootfs=self.devrootfs)
 		else:
 			self.close()
 
 	def green(self):
 		if self.check_hdd():
-			self.session.open(doFlashImage, online = True, list=self.list[self.selection], multi=self.multi, devrootfs=self.devrootfs)
+			self.session.open(doFlashImage, online = True, list=self.list[self.selection], multi=self.multi, AZrootfsMTD = self.AZrootfsMTD, AZkernelMTD = self.AZkernelMTD, devrootfs=self.devrootfs)
 		else:
 			self.close()
 
 	def yellow(self):
-		if SystemInfo["HaveMultiBoot"]:
+		if SystemInfo["HaveMultiBoot"] and not getBoxType().startswith('azboxm'):
 			self.selection = self.selection + 1
 			if self.selection == len(self.list):
 				self.selection = 0
@@ -171,7 +195,7 @@ class FlashOnline(Screen):
 
 	def list_files(self, PATH):
 		files = []
-		if SystemInfo["HaveMultiBoot"]:
+		if SystemInfo["HaveMultiBoot"] and not getBoxType().startswith('azboxm'):
 			path = PATH
 			for name in os.listdir(path):
 				if name != 'bootname' and os.path.isfile(os.path.join(path, name)):
@@ -201,7 +225,7 @@ class doFlashImage(Screen):
 		<widget name="imageList" position="10,10" zPosition="1" size="450,450" font="Regular;20" scrollbarMode="showOnDemand" transparent="1" />
 	</screen>"""
 		
-	def __init__(self, session, online, list=None, multi=None, devrootfs=None ):
+	def __init__(self, session, online, list=None, multi=None, AZrootfsMTD=None, AZkernelMTD=None, devrootfs=None ):
 		Screen.__init__(self, session)
 		self.session = session
 
@@ -216,6 +240,10 @@ class doFlashImage(Screen):
 		self.Online = online
 		self.List = list
 		self.multi=multi
+		self.AZrootfsMTD = AZrootfsMTD
+		self.AZkernelMTD = AZkernelMTD
+		self.AZFlashKernel = False
+		self.AZfailboot = None
 		self.devrootfs=devrootfs
 		self.imagePath = imagePath
 		self.feedurl = feedurl_atv
@@ -236,6 +264,17 @@ class doFlashImage(Screen):
 		self.newfeed = None
 		if os.path.exists('/etc/enigma2/newfeed'):
 			self.newfeed = ReadNewfeed()
+
+	def get_current_azbootnum(self):
+		if not os.path.isfile('/boot/.boot'):
+			os.system("mount -t jffs2 mtd2 /boot/")
+		if os.path.isfile('/boot/.boot'):
+			f=open('/boot/.boot','r')
+			current_bootnum = int(f.read())
+			f.close()		
+		#os.system("umount -fl /boot")
+		print "[Flash_online: current Azboot=]", current_bootnum
+		return current_bootnum
 
 	def quit(self):
 		if self.simulate or not self.List == "STARTUP":
@@ -366,7 +405,10 @@ class doFlashImage(Screen):
 			return
 		if len(job_manager.failed_jobs) == 0:
 			self.flashWithPostFlashActionMode = 'online'
-			self.flashWithPostFlashAction()
+			if self.box().startswith('azboxm'):
+				self.AZFlashOptionsAction()
+			else:
+				self.flashWithPostFlashAction()
 		else:
 			self.session.open(MessageBox, _("Download Failed !!"), type = MessageBox.TYPE_ERROR)
 
@@ -454,10 +496,76 @@ class doFlashImage(Screen):
 				else:
 					self.startInstallLocalCB()
 			else:
+				if self.box().startswith('azboxm'):
+					cmd = "echo "+str(self.AZfailboot)+" > /boot/.boot"
+					os.system(cmd)
 				self.show()
 		else:
 			self.show()
 
+	def AZFlashOptionsAction(self, ret = True):
+		if ret:
+			print "AZFlashOptionsAction"
+			title =_("Please select the flashing target.\n(When flashing onto different boot partition, the system will reboot into the target partition)")
+			list = ((_("Flash Kernel only"), "kernelonly"),
+			(_("Flash Kernel and Image to Boot 0"), "kernelrootfs0"),
+			(_("Flash Kernel and Image to Boot 1"), "kernelrootfs1"),
+			(_("Flash Kernel and Image to Boot 2"), "kernelrootfs2"),
+			(_("Flash 'Image only' to Boot 0"), "rootfs0"),
+			(_("Flash 'Image only' to Boot 1"), "rootfs1"),
+			(_("Flash 'Image only' to Boot 2"), "rootfs2"),
+			(_("Do not flash image"), "abort"))
+			self.session.openWithCallback(self.AZFlashActionCallback, ChoiceBox,title=title,list=list,selection=self.AZSelectedFlashOption())
+		else:
+			self.show()
+
+	def AZSelectedFlashOption(self):
+		index = 0
+		AZCurrentPartion = self.get_current_azbootnum()
+		if not self.AZfailboot:
+			self.AZfailboot = AZCurrentPartion
+		
+		if 	AZCurrentPartion == 0:
+			index = 1
+		elif AZCurrentPartion == 2:
+			index = 2
+		elif AZCurrentPartion == 3:
+			index = 3
+
+		return index
+
+	def AZFlashActionCallback(self, AZanswer):
+		print "AZFlashActionCallback"
+		if AZanswer is not None:
+			if AZanswer[1]== "kernelonly" or AZanswer[1] == "kernelrootfs0" or AZanswer[1] == "kernelrootfs1" or AZanswer[1] == "kernelrootfs2":
+			## TESTING IF THE FLASHTOOL FOR AZBOX KERNEL FLASHING IS PRESENT
+				if not os.path.exists("/usr/sbin/flashkernel"):
+					text = "Flashkernel not found!! Select one of the 'Flash Image only' options or Abort"
+					self.session.open(MessageBox, _(text), type = MessageBox.TYPE_ERROR, timeout = 10)
+					return False
+				self.AZFlashKernel = True
+			elif AZanswer[1] == "kernelrootfs0" or AZanswer[1] == "rootfs0":
+				self.AZkernelMTD = 2
+				self.AZrootfsMTD = self.AZkernelMTD + 1
+				os.system("echo 0 > /boot/.boot")
+			elif AZanswer[1] == "kernelrootfs1" or AZanswer[1] == "rootfs1":
+				self.AZkernelMTD = 4
+				self.AZrootfsMTD = self.AZkernelMTD + 1
+				os.system("echo 2 > /boot/.boot")
+			elif AZanswer[1] == "kernelrootfs2" or AZanswer[1] == "rootfs2":
+				self.AZkernelMTD = 6
+				self.AZrootfsMTD = self.AZkernelMTD + 1
+				os.system("echo 3 > /boot/.boot")
+#			os.system("umount -fl /boot")
+			if AZanswer[1] != "abort":
+				self.flashWithPostFlashAction()
+			else:
+				cmd = "echo "+str(self.AZfailboot)+" > /boot/.boot"
+				os.system(cmd)
+				self.show()
+		else:
+			self.show()
+			
 	def unzip_image(self, filename, path):
 		print "Unzip %s to %s" %(filename,path)
 		self.session.openWithCallback(self.cmdFinished, Console, title = _("Unzipping files, Please wait ..."), cmdlist = ['unzip ' + filename + ' -o -d ' + path, "sleep 3"], closeOnSuccess = True)
@@ -474,7 +582,12 @@ class doFlashImage(Screen):
 			if self.simulate:
 				text += _("Simulate (no write)")
 				if SystemInfo["HaveMultiBoot"]:
-					cmdlist.append("%s -n -r -k -m%s %s > /dev/null 2>&1" % (ofgwritePath, self.multi, flashTmp))
+					if not self.box().startswith('azboxm'):
+						cmdlist.append("%s -n -r -k -m%s %s > /dev/null 2>&1" % (ofgwritePath, self.multi, flashTmp))
+					elif self.AZFlashKernel:
+						cmdlist.append("%s -n -rmtd%s -kmtd%s %s > /dev/null 2>&1" % (ofgwritePath, self.AZrootfsMTD, self.AZkernelMTD, flashTmp))
+					else:
+						cmdlist.append("%s -n -rmtd%s %s > /dev/null 2>&1" % (ofgwritePath, self.AZrootfsMTD, flashTmp))
 				else:
 					cmdlist.append("%s -n -r -k %s > /dev/null 2>&1" % (ofgwritePath, flashTmp))
 				self.close()
@@ -484,16 +597,21 @@ class doFlashImage(Screen):
 			else:
 				text += _("root and kernel")
 				if SystemInfo["HaveMultiBoot"]:
-					if not self.List == "STARTUP":
-						os.system('mkfs.ext4 -F ' + self.devrootfs)
-					cmdlist.append("%s -r -k -m%s %s > /dev/null 2>&1" % (ofgwritePath, self.multi, flashTmp))
-					if not self.List == "STARTUP":
-						cmdlist.append("umount -fl /oldroot_bind")
-						cmdlist.append("umount -fl /newroot")
+					if not self.box().startswith('azboxm'):
+						if not self.List == "STARTUP":
+							os.system('mkfs.ext4 -F ' + self.devrootfs)
+						cmdlist.append("%s -r -k -m%s %s > /dev/null 2>&1" % (ofgwritePath, self.multi, flashTmp))
+						if not self.List == "STARTUP":
+							cmdlist.append("umount -fl /oldroot_bind")
+							cmdlist.append("umount -fl /newroot")
+					elif self.AZFlashKernel:
+						cmdlist.append("%s -rmtd%s -kmtd%s %s > /dev/null 2>&1" % (ofgwritePath, self.AZrootfsMTD, self.AZkernelMTD, flashTmp))
+					else:
+						cmdlist.append("%s -rmtd%s %s > /dev/null 2>&1" % (ofgwritePath, self.AZrootfsMTD, flashTmp))
 				else:
 					cmdlist.append("%s -r -k %s > /dev/null 2>&1" % (ofgwritePath, flashTmp))
 				message = "echo -e '\n"
-				if not self.List == "STARTUP" and SystemInfo["HaveMultiBoot"]:
+				if not self.List == "STARTUP" and SystemInfo["HaveMultiBoot"] and not self.box().startswith('azboxm'):
 					message += _('ofgwrite flashing ready.\n')
 					message += _('please press exit to go back to the menu.\n')
 				else:
@@ -528,7 +646,7 @@ class doFlashImage(Screen):
 					dest = flashTmp + '/%s' %KERNELBIN
 					shutil.copyfile(binfile, dest)
 					kernel = False
-				elif name.find('root') > -1 and (name.endswith('.bin') or name.endswith('.jffs2') or name.endswith('.bz2')) and rootfs:
+				elif (name.find('root') > -1 or name.find('image0') > -1) and (name.endswith('.bin') or name.endswith('.jffs2') or name.endswith('.bz2')) and rootfs:
 					binfile = os.path.join(path, name)
 					dest = flashTmp + '/%s' %ROOTFSBIN
 					shutil.copyfile(binfile, dest)
@@ -543,6 +661,11 @@ class doFlashImage(Screen):
 					dest = flashTmp + '/e2jffs2.img'
 					shutil.copyfile(binfile, dest)
 					rootfs = False
+				elif name.find('zbimage-linux-xload') > -1 and kernel:
+					binfile = os.path.join(path, name)
+					dest = flashTmp + '/zbimage-linux-xload'
+					shutil.copyfile(binfile, dest)
+					kernel = False
 					
 	def yellow(self):
 		if not self.Online:
@@ -554,11 +677,17 @@ class doFlashImage(Screen):
 		if ret:
 			from Plugins.SystemPlugins.SoftwareManager.BackupRestore import BackupScreen
 			self.flashWithPostFlashActionMode = 'local'
-			self.session.openWithCallback(self.flashWithPostFlashAction,BackupScreen, runBackup = True)
+			if self.box().startswith('azboxm'):
+				self.session.openWithCallback(self.AZFlashOptionsAction,BackupScreen, runBackup = True)
+			else:
+				self.session.openWithCallback(self.flashWithPostFlashAction,BackupScreen, runBackup = True)
 		else:
 			self.flashWithPostFlashActionMode = 'local'
-			self.flashWithPostFlashAction()
-
+			if self.box().startswith('azboxm'):
+				self.AZFlashOptionsAction()
+			else:
+				self.flashWithPostFlashAction()
+			
 	def startInstallLocalCB(self, ret = None):
 		if self.sel == str(flashTmp):
 			self.Start_Flashing()
@@ -637,7 +766,7 @@ class doFlashImage(Screen):
 			lines = the_page.split('\n')
 			tt = len(box)
 			for line in lines:
-				if line.find("<a href='%s/" % box) > -1:
+				if line.find("<a href='%s/" % box) > -1 and line.find("usb") > -1:
 					t = line.find("<a href='%s/" % box)
 					if self.feed == "atv" or self.feed == "atv2":
 						self.imagelist.append(line[t+tt+10:t+tt+tt+39])
